Fix error when previewing repeater fields ...
See also https://github.com/bolt/bolt/issues/6605

In preview mode, repeaters are arrays, whereas on saved pages
these are instances of `RepeatingFieldCollection`. For now, we can
supress the Bolt errors by checking the route.

An if-else statement _could_ be done, if we want to warn the users
explicitly. Alternatively, we need to re-write all the `.get(...)`
calls.